feat: Add resume
Prior to this change, the resume used to be required to be uploaded to
Contentful in pdf format every time an ammendment was made to the docx
file. The file name would have been changed, so the link would also
have to be updated in the markdown text in Contentful. Then the deploy
of the website would have to be manually triggered every time from
Netlify for every time the resume changes.

This change puts the resume directly in the GitHub repository. This
file would be linked to in the markdown from now on. The name of the
file will always be `ahadc-resume.pdf`, so the URL to this file will
always be the same. Therefore, no more updates to markdown for every
update in the resume, and no more manual triggering the deployment,
since it is automatically triggered every time a new commit is pushed!