fix(deploy): Remove Writing section temporarily
Prior to this change, there was a 'Writing' section that depended on a
package, `gatsby-source-medium`. That package works no longer due to a
recent change in Medium's API policy. For this, builds on Netlify were
failing with every commit, and no change was going live.

This change removes the section entirely, along with the package and the
plugin in `gatsby-config`. Now there is no problem anymore with builds
on Netlify.

This is temporary, and I am currently working on getting a workaround
for Medium posts. Might have to push the Medium data on Contentful and
pull from there via GraphQL.