chore(deps): Add Node.js version; upgrade packages
Prior to this change, Node.js version was not defined in `package.json`,
for which deployment of staging on Heroku was failing.

This change specifies the Node.js version the runtime should use.

This change also upgrades a lot of Node.js packages, since a few of them
were not compatible with Node.js v12.13.0.