fix(deploy): Fix "engines" in `package.json`
Prior to this change, the "engines" field was not correctly defined,
which was causing operations like 'yarn', causing issues in CI/CD.

This change fixes the issue. I took a look at the release cycle of
Node.js and figured out that v12 would be the best option for both
longevity (it is an even numbered version, and will be in active LTS
throughout the next 12 months) and novelty of features.

That awkward moment while you are writing a commit message you realize
that the commit message is about times 11 times longer than the length
of the actual commit character count 😂